pyproject.toml
[tool.poetry]
name = "athena"
version = "0.1.0"
description = "AI-powered infrastructure orchestration CLI"
authors = ["Athena Team"]
readme = "README.md"
packages = [{include = "athena_ai"}]

[tool.poetry.dependencies]
python = "^3.11"
# CLI & UI
click = "^8.1"
rich = "^14.0"
prompt_toolkit = "^3.0"
# Remote execution
paramiko = "^3.5"
fabric = "^3.2"
ansible-runner = "^2.4"
netmiko = "^4.4"
# AI/LLM providers (stable 2025 versions)
anthropic = "^0.70"
openai = "^2.0"
litellm = "^1.70"
openai-agents = "^0.6.0"
pyautogen = "^0.10.0"
# Infrastructure
python-terraform = "^0.10"
redis = "^5.2"
sqlalchemy = "^2.0"
# Data validation & security
pydantic = "^2.10"
cryptography = "^44.0"
keyring = "^25.5"
# Utilities
tenacity = "^9.0"
python-dotenv = "^1.0"
pyyaml = "^6.0"
jinja2 = "^3.1"
loguru = "^0.7"
# Data source connectors
psycopg2-binary = "^2.9"
pymysql = "^1.1"
pymongo = "^4.10"
requests = "^2.32"
# Knowledge system (optional)
falkordb = {version = "^1.2", optional = true}
duckduckgo-search = {version = "^8.0", optional = true}

[tool.poetry.extras]
knowledge = ["falkordb", "duckduckgo-search"]
all = ["falkordb", "duckduckgo-search"]

[tool.poetry.scripts]
athena = "athena_ai.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
